+++ client/src/components/SchoolCreate.js
